# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Transport Dispatcher API",
    "description": "NDTP and local historical CSV telemetry, scheduled arrivals, ML forecasts and alerts.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tracks": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Tracks",
        "description": "Recent live packets for restoring paths after a browser reload.",
        "operationId": "tracks_api_v1_tracks_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ndtp",
                    "api"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrackPoint"
                  },
                  "title": "Response Tracks Api V1 Tracks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stops": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Stops",
        "operationId": "stops_api_v1_stops_get",
        "parameters": [
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At"
            }
          },
          {
            "name": "before_min",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 120,
              "minimum": 0,
              "default": 0,
              "title": "Before Min"
            }
          },
          {
            "name": "after_min",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 120,
              "minimum": 0,
              "default": 60,
              "title": "After Min"
            }
          },
          {
            "name": "tr_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tr Id"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "live",
                "historical"
              ],
              "type": "string",
              "default": "live",
              "title": "Mode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduledArrival"
                  },
                  "title": "Response Stops Api V1 Stops Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/registrations": {
      "get": {
        "tags": [
          "setup"
        ],
        "summary": "Registrations",
        "operationId": "registrations_api_v1_registrations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VehicleRegistration"
                  },
                  "type": "array",
                  "title": "Response Registrations Api V1 Registrations Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Register",
        "description": "Map an NDTP unit_id to the transport ID used in the schedule.",
        "operationId": "register_api_v1_registrations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleRegistration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleRegistration"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/arrivals": {
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Scheduled Arrival",
        "description": "Add or update one planned arrival; required for selecting a 10–15 minute target.",
        "operationId": "scheduled_arrival_api_v1_arrivals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduledArrival"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledArrival"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/arrivals/{arrival_id}/observed": {
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Observed Arrival",
        "description": "Record a completed stop so the ML input includes current_dev_s.",
        "operationId": "observed_arrival_api_v1_arrivals__arrival_id__observed_post",
        "parameters": [
          {
            "name": "arrival_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Arrival Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObservedArrival"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArrivalDeviation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/snapshot": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Snapshot",
        "operationId": "snapshot_api_v1_snapshot_get",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "live",
                "historical"
              ],
              "type": "string",
              "default": "live",
              "title": "Mode"
            }
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vehicles": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Vehicles",
        "operationId": "vehicles_api_v1_vehicles_get",
        "parameters": [
          {
            "name": "risk",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(green|yellow|red|unknown)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "live",
                "historical"
              ],
              "type": "string",
              "default": "live",
              "title": "Mode"
            }
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Vehicle"
                  },
                  "title": "Response Vehicles Api V1 Vehicles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vehicles/{tr_id}": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Vehicle",
        "operationId": "vehicle_api_v1_vehicles__tr_id__get",
        "parameters": [
          {
            "name": "tr_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tr Id"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "live",
                "historical"
              ],
              "type": "string",
              "default": "live",
              "title": "Mode"
            }
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Alerts",
        "operationId": "alerts_api_v1_alerts_get",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "live",
                "historical"
              ],
              "type": "string",
              "default": "live",
              "title": "Mode"
            }
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Alert"
                  },
                  "title": "Response Alerts Api V1 Alerts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TrackPoint": {
        "properties": {
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id"
          },
          "tr_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tr Id"
          },
          "unit_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Id"
          },
          "event_time": {
            "type": "string",
            "format": "date-time",
            "title": "Event Time"
          },
          "lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "speed_kmh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Kmh"
          },
          "location_valid": {
            "type": "boolean",
            "title": "Location Valid"
          },
          "source": {
            "type": "string",
            "enum": [
              "ndtp",
              "api",
              "csv"
            ],
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "event_time",
          "location_valid",
          "source"
        ],
        "title": "TrackPoint"
      },
...
    }
  }
}